float

+++ code/render_loss.py
@@ -11,9 +11,9 @@ f.close()
 
 x_array = np.arange(len(contents_dis))
 
-contents_gen = np.array(contents_gen)
+contents_gen = np.array(contents_gen, dtype=np.float32)
 plt.plot(x_array, contents_gen, color='blue')
 
-contents_dis = np.array(contents_dis)
+contents_dis = np.array(contents_dis, dtype=np.float32)
 plt.plot(x_array, contents_dis, color='green')
 plt.savefig('loss_plot.png')
